requirements.txt
Flask
pandas
numpy
scikit-learn
joblib
python-dotenv
gunicorn
google-genai

